[SWM-161] test: DailyHistoryResponseDto fixture type fix

+++ src/test/java/com/climbx/climbx/fixture/UserFixture.java
@@ -183,7 +183,7 @@ public class UserFixture {
     ) {
         return DailyHistoryResponseDto.builder()
             .date(date)
-            .value(value)
+            .value((long) value)
             .build();
     }
